requirements.txt
flask
boto3
python-dotenv
flask-cors
requests
flask-session

